% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{build_catalogue}
\alias{build_catalogue}
\title{Builds a mutational catalogue from a table containing the base change and
trinucleotide context of each single-nucleotide variant.}
\usage{
build_catalogue(variants)
}
\arguments{
\item{variants}{Table with one row per single-nucleotide variant, and three columns:
REF base (character in {A,C,G,T}); ALT base (character in {A,C,G,T}); and trinucleotide
context at the variant location (sequence between the positions immediately before and 
after the variant, in string format; e.g. "TCA").}
}
\description{
Builds a mutational catalogue from a table containing the base change and
trinucleotide context of each single-nucleotide variant.
}
