% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{calculate_gof}
\alias{calculate_gof}
\title{Do goodness-of-fit calculations}
\usage{
calculate_gof(sample_list, stat = "cosine")
}
\arguments{
\item{sample_list}{List containing the results of signature extraction using
\code{\link{extract_signatures}} with multiple numbers of signatures (see argument
\code{nsignatures} in \code{\link{extract_signatures}}).}

\item{stat}{Function for measuring goodness-of-fit. Admits character values \code{"cosine"}
(cosine similarity, default) or \code{"L2"} (L2 norm or Euclidean distance).}
}
\value{
List of nS, number of signatures; gof, goodness of fit score; best, index of best
scoring signature; stat, the statistic used to calculate goodness of fit (cosine or L2); model,
sigfit model name obtained from the samples.
}
\description{
\code{calculate_gof} Calculates the goodness-of-fit information for a set of samples, each of which has typically been
sampled using an signature extraction model with a different number of signatures.
}
