% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{convert_signatures}
\alias{convert_signatures}
\title{Convert signatures between models}
\usage{
convert_signatures(signatures, ref_opportunities, model_to)
}
\arguments{
\item{signatures}{Either a numeric matrix of mutational signatures, with one row per signature
and one column per mutation type, or a list of matrices generated via \code{\link{retrieve_pars}}.}

\item{ref_opportunities}{Numeric vector of reference or average mutational opportunities, with
one element per mutation type. It also admits character values \code{"human-genome"} or
\code{"human-exome"}, in which case the mutational opportunities of the reference human
genome/exome will be used.}

\item{model_to}{Model to convert to; admits character values \code{"nmf"} (in which case the
signatures will be multiplied by the opportunities) or \code{"emu"} (in which case the signatures
will be divided by the opportunities).}
}
\value{
A numeric matrix of transformed signatures with the same dimensions as \code{signatures}.
}
\description{
\code{convert_signatures} converts between a representation of signatures relative to the
reference mutational opportunities (such as used in conventional NMF approaches), and a
representation which is not relative to the mutational opportunities (as used in the EMu software).
This is done by multiplying or dividing each signature by the average mutational opportunities
of the samples, or by the human genome/exome reference trinucleotide frequencies.
}
\examples{
# Load COSMIC signatures
# These are in "NMF" format, i.e. they are relative
# to the human genome mutational opportunities
data("cosmic_signatures_v2")

# Plot COSMIC signature 1
barplot(cosmic_signatures_v2[1,])

# Convert signatures to the "EMu" format, i.e. make
# them not relative to mutational opportunities
converted_signatures <- convert_signatures(cosmic_signatures_v2,
                                           ref_opportunities = "human-genome",
                                           model_to = "emu")

# Plot COSMIC signature 1, converted to "EMu" format
barplot(converted_signatures[1,])
}
