% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosmic_signatures_indel_v3.2}
\alias{cosmic_signatures_indel_v3.2}
\title{Small insertion/deletion signatures from COSMIC v3.2}
\format{An object of class \code{matrix} with 18 rows and 83 columns.}
\source{
\url{https://cancer.sanger.ac.uk/cosmic/signatures/ID/}

\url{https://cancer.sanger.ac.uk/signatures/downloads/}
}
\usage{
cosmic_signatures_indel_v3.2
}
\description{
A data set of 18 ID signatures from COSMIC v3.2 (March 2021).
With a few exceptions, these signatures were extracted
using SigProfiler (as described in Alexandrov et al., 2020) from the 2,780
whole-genome variant calls produced by the ICGC/TCGA Pan Cancer Analysis of
Whole Genomes (PCAWG) Network.
}
\keyword{datasets}
