% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosmic_signatures_v2}
\alias{cosmic_signatures_v2}
\title{Single base substitution signatures from COSMIC v2}
\format{An object of class \code{matrix} with 30 rows and 96 columns.}
\source{
\url{https://cancer.sanger.ac.uk/cosmic/signatures/SBS/}

\url{https://cancer.sanger.ac.uk/signatures/downloads/}
}
\usage{
cosmic_signatures_v2
}
\description{
A data set of 30 SBS signatures from COSMIC v2 (March 2015).
These signatures were extracted using the WTSI Mutational Signature Framework
as described in Alexandrov et al., 2013.
}
\keyword{datasets}
