% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosmic_signatures_v3_strand}
\alias{cosmic_signatures_v3_strand}
\title{Single base substitution signatures from COSMIC v3 with strand bias}
\format{An object of class \code{matrix} with 65 rows and 192 columns.}
\source{
\url{https://cancer.sanger.ac.uk/cosmic/signatures/SBS/}

\url{https://www.synapse.org/#!Synapse:syn11801889}
}
\usage{
cosmic_signatures_v3_strand
}
\description{
A data set of 67 strand-bias signatures from COSMIC v3.0 (May 2019).
These signatures were extracted using SigProfiler (as described in Alexandrov
et al., 2020) from the 2,780 whole-genome variant calls produced by the
ICGC/TCGA Pan Cancer Analysis of Whole Genomes (PCAWG) Network.
The following signatures are annotated as "Possible sequencing artefacts":
SBS27, SBS43, SBS45-60.
}
\keyword{datasets}
