% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{default_sig_palette}
\alias{default_sig_palette}
\title{Generate default colour palette for signatures}
\usage{
default_sig_palette(n)
}
\description{
Generate default colour palette for signatures
}
