% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigplot.R
\name{exposure_dataframe}
\alias{exposure_dataframe}
\title{Extract exposures data from extracted rstan samples array}
\usage{
exposure_dataframe(exp)
}
\arguments{
\item{exp}{item such as stanfit_extract$signatures[, 1, ]}
}
\description{
Extract exposures data from extracted rstan samples array
}
