% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{extract_signatures}
\alias{extract_signatures}
\title{Extract signatures from a set of mutation counts}
\usage{
extract_signatures(counts, nsignatures, method = "emu",
  opportunities = NULL, sig_prior = NULL, stanfunc = "sampling", ...)
}
\arguments{
\item{counts}{Matrix of mutation counts for each sample (rows) in each category
(columns).}

\item{nsignatures}{Number (or range of numbers) of signatures to extract.}

\item{method}{Either \code{"emu"} (default) or \code{"nmf"}.}

\item{opportunities}{Optional matrix of mutational opportunities for \code{"emu"} method; 
must have same dimension as \code{counts}. If equals to \code{"human-genome"} or 
\code{"human-exome"}, the reference human genome/exome opportunities will be used for every sample.}

\item{sig_prior}{Matrix with one row per signature and one column per category, to be used as the Dirichlet 
priors for the signatures to be extracted. Default priors are uniform (uninformative).}

\item{stanfunc}{\code{"sampling"}|\code{"optimizing"}|\code{"vb"} Choice of rstan inference strategy. 
\code{"sampling"} is the full Bayesian MCMC approach, and is the default. \code{"optimizing"}
returns the Maximum a Posteriori (MAP) point estimates via numerical optimization.
\code{"vb"} uses Variational Bayes to approximate the full posterior.}

\item{...}{Any other parameters to pass to rstan.}
}
\description{
Extract signatures from a set of mutation counts
}
