% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_estimation.R
\name{extract_signatures_initialiser}
\alias{extract_signatures_initialiser}
\title{Use optimization to generate initial parameter values for MCMC sampling}
\usage{
extract_signatures_initialiser(counts, nsignatures, method = "emu",
  opportunities = NULL, sig_prior = NULL, chains = 1, ...)
}
\description{
Use optimization to generate initial parameter values for MCMC sampling
}
