% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{extract_signatures_opt_init}
\alias{extract_signatures_opt_init}
\title{Use optimization to generate initial parameter values for MCMC sampling}
\usage{
extract_signatures_opt_init(counts, nsignatures, method = "emu",
  opportunities = NULL, sig_prior = NULL, ...)
}
\description{
Use optimization to generate initial parameter values for MCMC sampling
}
