% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{get_loglik}
\alias{get_loglik}
\title{Generate log likelihood values from a model}
\usage{
get_loglik(counts, mcmc_samples)
}
\arguments{
\item{counts}{Matrix of observed mutation counts (integers), with one row per sample and 
column for each of the 96 mutation types.}

\item{mcmc_samples}{Object of class stanfit, generated via either \code{\link{fit_signatures}}
or \code{\link{extract_signatures}}.}
}
\description{
Generate log likelihood values from a model
}
