% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{get_reconstructions}
\alias{get_reconstructions}
\title{Generate reconstructed mutation catalogues from parameters estimated from MCMC samples}
\usage{
get_reconstructions(counts, mcmc_samples, signatures = NULL)
}
\arguments{
\item{counts}{Matrix of observed mutation counts (integers), with one row per sample and 
column for each of the 96 mutation types.}

\item{mcmc_samples}{Object of class stanfit, generated via either \code{\link{fit_signatures}}
or \code{\link{extract_signatures}}.}

\item{signatures}{Matrix of signatures to use for fitting models, which make no estimate of signatures.}
}
\description{
Generate reconstructed mutation catalogues from parameters estimated from MCMC samples
}
