% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigplot.R
\name{ggplot_bic}
\alias{ggplot_bic}
\title{Plot the BIC values as violin plots}
\usage{
ggplot_bic(sample_list)
}
\arguments{
\item{sample_list}{- list of stanfit samples produced by extract_signatures for a range of numbers of signatures}
}
\description{
Plot the BIC values as violin plots
}
