% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigplot.R
\name{ggplot_reconstruction}
\alias{ggplot_reconstruction}
\title{Returns a list of normalised exposure plots, with error bars.}
\usage{
ggplot_reconstruction(samples, counts, opportunities = NULL)
}
\arguments{
\item{samples}{Sampling object such as returned by extract_signatures.}
}
\description{
Returns a list of normalised exposure plots, with error bars.
}
\examples{
Generate a list of exposure plots for all samples, showing 3 extracted signatures
samples <- extract_signatures(mycounts, nsignatures = 3, method = "nmf") 
plots <- plot_exposures(samples)
m <- marrangeGrob(plots, nrow = 1, ncol = 3) # use gridExtra to plot all on same page
}
