% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigplot.R
\name{ggplot_signatures}
\alias{ggplot_signatures}
\title{Returns a list of signature spectrum plots, with error bars.}
\usage{
ggplot_signatures(samples)
}
\arguments{
\item{samples}{Sampling object such as returned by extract_signatures.}
}
\description{
Returns a list of signature spectrum plots, with error bars.
}
\examples{
Generate a list of 3 plots for 3 extracted signatures
samples <- extract_signatures(mycounts, nsignatures = 3, method = "nmf") 
plots <- plot_signatures(samples)
m <- marrangeGrob(plots, nrow = 1, ncol = 3) # use gridExtra to plot all on same page
}
