% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{human_trinuc_freqs}
\alias{human_trinuc_freqs}
\title{Returns human genome or exome trinucleotide frequencies. This is useful to
de-normalize signatures that were obtained using the mutational opportunities
from the human genome/exome, in order to compare them to signatures extracted
without incorporating opportunities.}
\usage{
human_trinuc_freqs(type = "genome")
}
\arguments{
\item{type}{Either "genome" (default) or "exome".}
}
\description{
Returns human genome or exome trinucleotide frequencies. This is useful to
de-normalize signatures that were obtained using the mutational opportunities
from the human genome/exome, in order to compare them to signatures extracted
without incorporating opportunities.
}
\examples{
Extract signatures using human exome opportunitites
samples <- extract_signatures(mycounts, nsignatures = 3, method = "emu", 
opportunities = "human-exome")
sigs <- retrieve_pars(samples, "signatures")

# De-normalize (mean) extracted signatures
freqs <- human_trinuc_freqs("exome")
sigs_denorm <- apply(sigs, 1, function(sig) { 
     tmp <- sig[,1] * freqs
     tmp / sum(tmp)
})
}
