% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{letterwrap}
\alias{letterwrap}
\title{From https://stackoverflow.com/a/21689613/1875814
Produce Excel-style letter labels - A, B, ..., Z, AA, AB, ..., AZ, ..., ZZ, AAA, ...}
\usage{
letterwrap(n, depth = 1)
}
\description{
From https://stackoverflow.com/a/21689613/1875814
Produce Excel-style letter labels - A, B, ..., Z, AA, AB, ..., AZ, ..., ZZ, AAA, ...
}
