% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_plotting.R
\name{make_colors}
\alias{make_colors}
\title{Colour warnings for posterior predictive checks
yellow = observed value is in top or bottom 2.5% of ppc simulations
red = observed value is outside ppc simulations}
\usage{
make_colors(c, c_ppc, sample = 1)
}
\description{
Colour warnings for posterior predictive checks
yellow = observed value is in top or bottom 2.5% of ppc simulations
red = observed value is outside ppc simulations
}
