% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{mut_types}
\alias{mut_types}
\title{Returns character vector of 96 (pyrimidine) trinucleotide mutation types}
\usage{
mut_types()
}
\description{
Returns character vector of 96 (pyrimidine) trinucleotide mutation types
}
