% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigplot.R
\name{plot_bic_violin}
\alias{plot_bic_violin}
\title{Plot the BIC values of a list of stanfit objects as violin plots}
\usage{
plot_bic_violin(sample_list, fill = NULL, color = NULL)
}
\arguments{
\item{sample_list}{- list of stanfit samples, such as those produced 
by extract_signatures when given a range of numbers of signatures to 
extract}
}
\description{
Plot the BIC values of a list of stanfit objects as violin plots
}
