% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{plot_gof}
\alias{plot_gof}
\title{Plot goodness of fit}
\usage{
plot_gof(sample_list, counts, stat = "cosine")
}
\arguments{
\item{sample_list}{List of objects of class stanfit. Elements of the list
which are not of class stanfit are ignored.}

\item{counts}{Matrix of mutation counts, with 96 columns and one row per catalogue.}

\item{stat}{Function for measuring goodness of fit. Admits values "cosine" 
(cosine similarity; default) or "L2" (L2 norm, a.k.a. Euclidean distance).}
}
\description{
\code{plot_gof} plots the goodness of fit of a set of samples, each of which
has typically been sampled using an extraction model (EMu or NMF) with a 
different number of signatures.
}
