% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_plotting.R
\name{plot_ppc}
\alias{plot_ppc}
\title{Plot result of posterior predictive check}
\usage{
plot_ppc(c, c_ppc, sample = 1)
}
\arguments{
\item{c}{Integer matrix of observed counts.}

\item{c_ppc}{Simulated posterior predictive counts obtained from
MCMC samples using \code{extract(samples)$counts_ppc}.}
}
\description{
black = observed value is not extreme compared to the simulated distribution
yellow = observed value is in top or bottom 2.5% of ppc simulations
red = observed value is outside ppc simulations
}
