% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{plot_reconstruction}
\alias{plot_reconstruction}
\title{Plot mutational spectrum reconstructions}
\usage{
plot_reconstruction(counts, mcmc_samples = NULL, signatures = NULL,
  exposures = NULL, opportunities = NULL, pdf_path = NULL,
  sig_color_palette = NULL)
}
\arguments{
\item{counts}{Matrix of observed mutation counts, with 96 columns and
one row per sample.}

\item{mcmc_samples}{Object of class stanfit, produced by either \code{\link{fit_signatures}}
or \code{\link{extract_signatures}}. Needed if \code{counts}, \code{signatures} or \code{exposures}
are not provided.}

\item{signatures}{Either a matrix of mutational signatures, with 96 columns and one row per
signature, or a list of signatures as produced by \code{\link{retrieve_pars}}. Only needed if 
\code{mcmc_samples} is not provided.}

\item{exposures}{Either a matrix of signature exposures, with one row per sample and one column 
per signature, or a list of exposures as produced by \code{\link{retrieve_pars}}. Only needed if \code{mcmc_samples} is not provided.}

\item{opportunities}{If signatures or exposures were obtained using \code{method="emu"},
the opportunities used for extraction/fitting need to be provided here. Admits values \code{"human-genome"}
and \code{"human-exome"}. Only needed if \code{mcmc_samples} is not provided.}

\item{pdf_path}{If provided, the plots will be output to a PDF file with this path.}

\item{sig_color_palette}{Character vector of colour names or hexadecimal codes to use for each signature.
Must have at least as many elements as the number of signatures.}
}
\description{
\code{plot_reconstruction} plots the reconstructions of the original mutational catalogues using the 
signatures and/or exposures obtained through extraction or fitting. If provided with multiple catalogues, it generates one
plot per catalogue. Data can be provided either as a single stanfit object resulting from extraction/fitting, 
or as separate objects for mutation counts, signatures and exposures. The latter option does not allow
plotting of error bars in the reconstructed catalogue.
}
\examples{
# Extract signatures using the EMu (Poisson) model
samples <- extract_signatures(mycounts, nsignatures = 3, method = "emu", 
opportunities = "human-genome")

# Retrieve signatures and exposures
signatures <- retrieve_pars(samples, "signatures")
exposures <- retrieve_pars(samples, "exposures")

# Plot reconstructed catalogues
pdf("Reconstruction.pdf", width = 24, height = 18)
par(mar = c(8, 8, 7, 2.75))
plot_reconstruction(mycounts, signatures, exposures, opportunities = "human-genome")
dev.off()
}
