% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_plotting.R
\name{plot_spectrum_tsw}
\alias{plot_spectrum_tsw}
\title{Plot strandwise SBS spectrum}
\usage{
plot_spectrum_tsw(spec, lwr, upr, name, max_y, colors, boxes)
}
\description{
Plot strandwise SBS spectrum
}
