% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{retrieve_pars}
\alias{retrieve_pars}
\title{Obtains summary values for a set of model parameters (signatures or exposures) from a stanfit object.}
\usage{
retrieve_pars(object, feature, prob = 0.95, signature_names = NULL)
}
\arguments{
\item{object}{An object of class stanfit.}

\item{feature}{Name of the parameter set to extract; either "signatures" or "exposures".}

\item{prob}{A numeric scalar in the interval (0,1) giving the target probability content of the HPD intervals.}

\item{signature_names}{Vector containing the names of the signatures used for fitting. Used only when 
retrieving exposures from fitted signatures.}
}
\description{
Obtains summary values for a set of model parameters (signatures or exposures) from a stanfit object.
}
\examples{
# Extract signatures using the EMu (Poisson) model
samples <- extract_signatures(mycounts, nsignatures = 3, method = "emu", 
opportunities = "human-genome")

# Retrieve array of signatures
signatures <- retrieve_pars(samples, "signatures")

# Retrieve array of exposures using 90\% HPD intervals
exposures <- retrieve_pars(samples, "exposures", prob = 0.9)

# Plot mean exposures
barplot(exposures$mean)  ## or barplot(exposures[[1]])
}
