\name{sigfit-package}
\alias{sigfit-package}
\alias{sigfit}
\docType{package}
\title{
\packageTitle{sigfit}
}
\description{
\packageDescription{sigfit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sigfit}
\packageIndices{sigfit}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{sigfit}

Maintainer: \packageMaintainer{sigfit}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
