% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigplot.R
\name{signature_dataframe}
\alias{signature_dataframe}
\title{Extract signature data from extracted rstan samples array}
\usage{
signature_dataframe(sig)
}
\arguments{
\item{sig}{item such as stanfit_extract$signatures[, 1, ]}
}
\description{
Extract signature data from extracted rstan samples array
}
