% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigplot.R
\name{sigplot}
\alias{sigplot}
\title{Plots a mutational spectrum}
\usage{
sigplot(spectrum, titletext = "")
}
\arguments{
\item{spectrum}{Vector of mutation rates}

\item{titletext}{A string to use as the title of the plot}
}
\description{
Plots a mutational spectrum
}
