% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{simulate_ppc}
\alias{simulate_ppc}
\title{Generate posterior predictive check values from a model}
\usage{
simulate_ppc(counts, mcmc_samples, size)
}
\description{
Generate posterior predictive check values from a model
}
