% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{to_matrix}
\alias{to_matrix}
\title{Initial coercion to matrix for signatures/exposures/counts}
\usage{
to_matrix(x, int = FALSE)
}
\description{
Initial coercion to matrix for signatures/exposures/counts
}
