% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cosmic_signatures_v3_strand}
\alias{cosmic_signatures_v3_strand}
\title{Single base substitution signatures from COSMIC v3 with strand bias}
\format{An object of class \code{matrix} with 65 rows and 192 columns.}
\source{
\url{https://cancer.sanger.ac.uk/cosmic/signatures/SBS/}

\url{https://www.synapse.org/#!Synapse:syn11801889}
}
\usage{
cosmic_signatures_v3_strand
}
\description{
A data set of 67 strand bias signatures extracted by SigProfiler as described
in Alexandrov, Kim, Hardhvala, Huang, et al., 2018.
Signatures 51-67 are annotated as "Possible sequencing artefacts"
}
\keyword{datasets}
