% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_estimation.R
\name{extract_signatures}
\alias{extract_signatures}
\title{Extract signatures from a set of mutation counts}
\usage{
extract_signatures(counts, nsignatures, method = "emu",
  opportunities = NULL, sig_prior = NULL, stanfunc = "sampling",
  cores = "auto", ...)
}
\arguments{
\item{counts}{Matrix of observed mutation counts (integers), with one row per sample and 
column for each of the 96 mutation types.}

\item{nsignatures}{Number (or vector of numbers) of signatures to extract.}

\item{method}{Either \code{"emu"} (default) or \code{"nmf"}.}

\item{opportunities}{Optional matrix of mutational opportunities for the "EMu" model 
(\code{method = "emu"}) method. Must be a matrix with same dimension as \code{counts}. 
If equal to \code{"human-genome"} or \code{"human-exome"}, the reference human genome/exome 
opportunities will be used for every sample.}

\item{sig_prior}{Matrix with one row per signature and one column per category, to be used as the Dirichlet 
priors for the signatures to be extracted. Only used when \code{nsignatures} is a scalar.
Default priors are uniform (uninformative).}

\item{stanfunc}{Choice of rstan inference strategy; admits values \code{"sampling"}, \code{"optimizing"}
and \code{"vb"}. \code{"sampling"} is the full Bayesian MCMC approach, and is the default. 
\code{"optimizing"} returns the Maximum a Posteriori (MAP) point estimates via numerical optimization.
\code{"vb"} uses Variational Bayes to approximate the full posterior.}

\item{cores}{Positive integer, number of cores to use. The default (\code{cores="auto"}) assigns
as many cores as detected in the system by \code{\link{parallel::detectCores()}}.}

\item{...}{Any other parameters to pass to the sampling function (by default, \code{\link{rstan::sampling}}).
(The number of chains is set to 1 and cannot be changed, to prevent 'label switching' problems.)}
}
\value{
A stanfit object containing the Monte Carlo samples from MCMC (from which the model
parameters can be extracted using \code{\link{retrieve_parameters}}), as well as information about
the model and sampling process.
}
\description{
Extract signatures from a set of mutation counts
}
\examples{
# Load example mutational catalogues
data("counts_21breast")

# Extract 2 to 6 signatures using the NMF (multinomial) model
# (400 warmup iterations + 400 sampling iterations -- use more in practice)
samples_nmf <- extract_signatures(counts_21breast, nsignatures = 2:6, 
                                  method = "nmf", iter = 800)

# Extract 4 signatures using the EMu (Poisson) model
# (400 warmup iterations + 800 sampling iterations -- use more in practice)
samples_emu <- extract_signatures(counts_21breast, nsignatures = 4, method = "emu", 
                                  opportunities = "human-genome",
                                  iter = 1200, warmup = 400)
}
