% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{extract_signatures}
\alias{extract_signatures}
\title{Extracts signatures from a set of mutation counts
using models based on NMF or EMu}
\usage{
extract_signatures(counts, nsignatures, method = "emu",
  opportunities = NULL, exposures_prior = 0.5, stanfunc = "sampling", ...)
}
\arguments{
\item{counts}{Matrix of mutation counts for each sample (rows) in each category
(columns).}

\item{nsignatures}{Number (or range of numbers) of signatures to extract.}

\item{method}{Either "emu" or "nmf" (though currently "nmf" is experimental).}

\item{opportunities}{Optional matrix of mutational opportunities for "emu" method; must have same dimension as counts. 
If equals to "human-genome" or "human-exome", the reference human genome/exome opportunities will be used for every sample.}

\item{exposures_prior}{Single numeric value to use for all the exposure priors; by default, 0.5 (i.e. Jeffreys prior).}

\item{stanfunc}{"sampling"|"optimizing"|"vb" Choice of rstan inference strategy. 
"sampling" is the full Bayesian MCMC approach, and is the default. "optimizing"
returns the Maximum a Posteriori (MAP) point estimates via numerical optimization.
"vb" uses Variational Bayes to approximate the full posterior.}

\item{...}{Any other parameters to pass through to rstan.}
}
\description{
Extracts signatures from a set of mutation counts
using models based on NMF or EMu
}
