% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{extract_signatures}
\alias{extract_signatures}
\title{Extracts signatures from a set of mutation counts
using models based on NMF or EMu}
\usage{
extract_signatures(counts, nsignatures, method = "emu",
  opportunities = NULL, stanfunc = "sampling", ...)
}
\arguments{
\item{counts}{Matrix of mutation counts for each sample (rows) in each category
(columns)}

\item{nsignatures}{Number of signatures to extract}

\item{method}{Either "emu" or "nmf" (though currently "nmf" is disabled)}

\item{opportunities}{Optional matrix of mutational opportunities for "emu" method.
Dimensions should be same as for counts}

\item{stanfunc}{"sampling"|"optimizing"|"vb" Choice of rstan inference strategy. 
"sampling" is the full Bayesian MCMC approach, and is the default. "optimizing"
returns the Maximum a Posteriori (MAP) point estimates via numerical optimization.
"vb" uses Variational Bayes to approximate the full posterior.}

\item{...}{Any other parameters to pass through to rstan}
}
\description{
Extracts signatures from a set of mutation counts
using models based on NMF or EMu
}
