% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_estimation.R
\name{extract_signatures_initialiser}
\alias{extract_signatures_initialiser}
\title{Use optimization to generate initial parameter values for MCMC sampling}
\usage{
extract_signatures_initialiser(counts, nsignatures, model = "emu",
  opportunities = NULL, sig_prior = NULL, exp_prior = 1,
  chains = 1, ...)
}
\arguments{
\item{counts}{Integer matrix of observed mutation counts, with one row per sample and
one column per mutation type. Any decimal values will be rounded to integers.}

\item{nsignatures}{Integer or integer vector indicating the number(s) of signatures to extract.}

\item{model}{Character indicating the model to sample from. Admits values \code{"nmf"} or 
\code{"emu"} (the default).}

\item{opportunities}{Numeric matrix of optional mutational opportunities for the "EMu" model
(\code{model = "emu"}). Must be a matrix with same dimension as \code{counts}.
Alternatively, it also admits character values \code{"human-genome"} or \code{"human-exome"},
in which case the reference human genome/exome opportunities will be used for every sample.}

\item{sig_prior}{Numeric matrix with one row per signature and one column per mutation type, 
to be used as the Dirichlet priors for the signatures to be extracted. Only used when 
\code{nsignatures} is a scalar. Default priors are uniform (uninformative).}

\item{chains}{Integer indicating number of chains to be initialised (default is 1).}

\item{...}{Additional arguments to pass to \code{rstan::optimizing}.}
}
\value{
List of initial values to be passed to \code{\link{extract_signatures}} via the
\code{init} argument.
}
\description{
Use optimization to generate initial parameter values for MCMC sampling
}
