% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_estimation.R
\name{extract_signatures_initialiser}
\alias{extract_signatures_initialiser}
\title{Use optimization to generate initial parameter values for MCMC sampling}
\usage{
extract_signatures_initialiser(counts, nsignatures,
  model = "multinomial", opportunities = NULL, sig_prior = NULL,
  exp_prior = NULL, chains = 1, ...)
}
\arguments{
\item{counts}{Numeric matrix of observed mutation counts, with one row per sample and
one column per mutation type.}

\item{nsignatures}{Integer or integer vector indicating the number(s) of signatures to extract.}

\item{model}{Name of the model to sample from. Admits character values \code{"multinomial"}
(default), \code{"poisson"}, \code{"negbin"}, \code{"normal"}, \code{"nmf"} (an alias for
\code{"multinomial"}), and \code{"emu"} (an alias for \code{"poisson"}).}

\item{opportunities}{Numeric matrix of optional mutational opportunities, with one row per sample
and one column per mutation type. It also admits character values \code{"human-genome"} or
\code{"human-exome"}, in which case the mutational opportunities of the reference human
genome/exome will be used for every sample.}

\item{sig_prior}{Numeric matrix with one row per signature and one column per mutation type,
to be used as the Dirichlet priors for the mutational signatures. Only used when a single value
is provided for \code{nsignatures}. Default priors are uniform.}

\item{chains}{Integer indicating number of chains to be initialised (default is 1).}

\item{...}{Additional arguments to be passed to \code{\link{rstan::optimizing}}.}
}
\value{
List of initial values to be passed to \code{\link{extract_signatures}} via the
\code{init} argument.
}
\description{
Use optimization to generate initial parameter values for MCMC sampling
}
