% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{fetch_cosmic_data}
\alias{fetch_cosmic_data}
\title{Fetch COSMIC mutational signatures}
\usage{
fetch_cosmic_data(reorder = TRUE, remove_zeros = TRUE)
}
\arguments{
\item{reorder}{If \code{TRUE}, the matrix will be reordered by substitution type and trinucleotide.}

\item{remove_zeros}{If \code{TRUE}, pseudocounts will be added to prevent the signatures from
containing any zeros, which can affect computation of the log likelihood.}
}
\value{
Matrix of signatures, with one row per signature and one column for each of
the 96 trinucleotide mutation types.
}
\description{
\code{fetch_cosmic_data} downloads the latest release of signatures from COSMIC
(http://cancer.sanger.ac.uk/cosmic/signatures) and produces a matrix of signatures
that can be used for signature fitting.
}
