% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{fit_extract_signatures}
\alias{fit_extract_signatures}
\title{Fit-and-extract mutational signatures}
\usage{
fit_extract_signatures(counts, signatures, num_extra_sigs, method = "nmf",
  opportunities = NULL, sig_prior = NULL, stanfunc = "sampling", ...)
}
\arguments{
\item{counts}{Matrix of mutation counts for each sample (rows) in each category
(columns).}

\item{signatures}{Matrix of fixed mutational signatures (columns) to be fitted.}

\item{num_extra_sigs}{Number of additional signatures to be extracted.}

\item{sig_prior}{Matrix with one row per additional signature and one column per category, to be used as the
Dirichlet priors for the additional signatures to be extracted. Default priors are uniform (uninformative).}

\item{stanfunc}{\code{"sampling"}|\code{"optimizing"}|\code{"vb"} Choice of rstan 
inference strategy. \code{"sampling"} is the full Bayesian MCMC approach, and is the 
default. \code{"optimizing"} returns the Maximum a Posteriori (MAP) point estimates 
via numerical optimization. \code{"vb"} uses Variational Bayes to approximate the 
full posterior.}

\item{...}{Any other parameters to pass through to rstan.}
}
\description{
\code{fit_extract_signatures} fits signatures to estimate exposures in a set of mutation counts
and extracts additional signatures present in the samples.
}
