% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_estimation.R
\name{fit_extract_signatures}
\alias{fit_extract_signatures}
\title{Fit-and-extract mutational signatures}
\usage{
fit_extract_signatures(counts, signatures, num_extra_sigs, method = "nmf",
  opportunities = NULL, sig_prior = NULL, stanfunc = "sampling", ...)
}
\arguments{
\item{counts}{Matrix of observed mutation counts (integers), with one row per sample and 
column for each of the 96 mutation types.}

\item{signatures}{Fixed mutational signatures (columns) to be fitted. Either a matrix with one row 
per signature and one column for each of the 96 mutation types, or a list of signatures generated via
\code{\link{retrieve_pars}}.}

\item{num_extra_sigs}{Number of additional signatures to be extracted.}

\item{sig_prior}{Matrix with one row per additional signature and one column per category, to be used as the
Dirichlet priors for the additional signatures to be extracted. Default priors are uniform (uninformative).}

\item{stanfunc}{\code{"sampling"}|\code{"optimizing"}|\code{"vb"} Choice of rstan 
inference strategy. \code{"sampling"} is the full Bayesian MCMC approach, and is the 
default. \code{"optimizing"} returns the Maximum a Posteriori (MAP) point estimates 
via numerical optimization. \code{"vb"} uses Variational Bayes to approximate the 
full posterior.}

\item{...}{Any other parameters to pass to the sampling function (by default, \code{\link{rstan::sampling}}).
(The number of chains is set to 1 and cannot be changed, to prevent 'label switching' problems.)}
}
\value{
A stanfit object containing the Monte Carlo samples from MCMC (from which the model
parameters can be extracted using \code{\link{retrieve_parameters}}), as well as information about
the model and sampling process.
}
\description{
\code{fit_extract_signatures} fits signatures to estimate exposures in a set of mutation counts
and extracts additional signatures present in the samples.
}
\examples{
# Fetch COSMIC signatures 
signatures <- fetch_cosmic_data()

# Simulate two catalogues using signatures 1, 4, 5, 7, with
# proportions 4:2:3:1 and 2:3:4:1, respectively
probs <- rbind(c(0.4, 0.2, 0.3, 0.1) \%*\% signatures[c(1, 4, 5, 7), ],
               c(0.2, 0.3, 0.4, 0.1) \%*\% signatures[c(1, 4, 5, 7), ])
mutations <- rbind(t(rmultinom(1, 20000, probs[1, ])),
                   t(rmultinom(1, 20000, probs[2, ])))  

# Assuming that we do not know signature 7 a priori, but we know the others
# to be present, extract 1 signature while fitting signatures 1, 4 and 5.
# (400 warmup iterations + 400 sampling iterations -- use more in practice)
mcmc_samples <- fit_extract_signatures(mutations, signatures = signatures[c(1, 4, 5), ],
                                       num_extra_sigs = 1, method = "nmf", iter = 800)

# Plot original and extracted signature 7
extr_sigs <- retrieve_pars(mcmc_samples, "signatures")
plot_spectrum(signatures[7, ], pdf_path = "COSMIC_Sig7.pdf", name="COSMIC sig. 7")
plot_spectrum(extr_sigs, pdf_path = "Extracted_Sigs.pdf")
}
