% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{fit_signatures}
\alias{fit_signatures}
\title{Runs MCMC to fit signatures and estimate exposures}
\usage{
fit_signatures(counts, signatures, prior = NULL, method = "nmf",
  opportunities = NULL, ...)
}
\arguments{
\item{counts}{Matrix of mutation counts per category (columns) per genome sample (rows).}

\item{signatures}{Matrix of mutational signatures (rows) to be fitted.}

\item{prior}{Vector of the same length as signatures, to be used as the Dirichlet prior in the sampling chain. Default prior is uniform (uninformative).}

\item{method}{Either "nmf" (default) or "emu".}

\item{opportunities}{Optional matrix of mutational opportunities for "emu" method; must have same dimension as counts. 
If equals to "human-genome" or "human-exome", the reference human genome/exome opportunities will be used for every sample.}

\item{...}{Arguments to pass to rstan::sampling.}
}
\description{
Runs MCMC to fit signatures and estimate exposures
}
\examples{
 # Custom prior favours signature 1 over 2, 3 and 4
samples <- fit_signatures(mycounts, mysignatures, prior = c(5, 1, 1, 1))

# Run a single chain for quite a long time
samples <- fit_signatures(mycounts, mysignatures, chains = 1, niter = 13000, warmup = 3000)
}
