% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{fit_signatures}
\alias{fit_signatures}
\title{Run MCMC to fit signatures and estimate exposures}
\usage{
fit_signatures(counts, signatures, exp_prior = NULL, method = "nmf",
  opportunities = NULL, ...)
}
\arguments{
\item{counts}{Matrix of mutation counts per category (columns) per genome sample (rows).}

\item{signatures}{Matrix of mutational signatures (rows) to be fitted.}

\item{exp_prior}{Vector with one element per signature, to be used as the Dirichlet prior for 
the signature exposures in the sampling chain. Default prior is uniform (uninformative).}

\item{method}{Model to sample from; either \code{"nmf"} or \code{"emu"}.}

\item{opportunities}{Optional matrix of mutational opportunities for \code{"emu"} method; must have same 
dimension as \code{counts}. If equals to \code{"human-genome"} or \code{"human-exome"}, the reference h
uman genome/exome opportunities will be used for every sample.}

\item{...}{Arguments to pass to \code{rstan::sampling}.}
}
\description{
Run MCMC to fit signatures and estimate exposures
}
\examples{
 # Custom prior favours signature 1 over 2, 3 and 4
samples <- fit_signatures(mycounts, mysignatures, exp_prior = c(5, 1, 1, 1))

# Run a single chain for quite a long time
samples <- fit_signatures(mycounts, mysignatures, chains = 1, niter = 13000, warmup = 3000)
}
