% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{fit_signatures}
\alias{fit_signatures}
\title{Fit mutational signatures}
\usage{
fit_signatures(counts, signatures, exp_prior = NULL, method = "nmf",
  opportunities = NULL, ...)
}
\arguments{
\item{counts}{Matrix of observed mutation counts (integers), with one row per sample and 
column for each of the 96 mutation types.}

\item{signatures}{Mutational signatures to be fitted. Either a matrix with one row per signature
and one column for each of the 96 mutation types, or a list of signatures generated via
\code{\link{retrieve_pars}}.}

\item{exp_prior}{Vector with one element per signature, to be used as the Dirichlet prior for 
the signature exposures in the sampling chain. Default prior is uniform (uninformative).}

\item{method}{Model to sample from; either \code{"nmf"} or \code{"emu"}.}

\item{opportunities}{Optional matrix of mutational opportunities for the "EMu" model 
(\code{method = "emu"}) method. Must be a matrix with same dimension as \code{counts}. 
If equal to \code{"human-genome"} or \code{"human-exome"}, the reference human genome/exome 
opportunities will be used for every sample.}

\item{...}{Arguments to pass to \code{rstan::sampling}.}
}
\value{
A stanfit object containing the Monte Carlo samples from MCMC (from which the model
parameters can be extracted using \code{\link{retrieve_parameters}}), as well as information about
the model and sampling process.
}
\description{
\code{fit_signatures} runs MCMC sampling to fit a set of mutational signatures 
to a collection of mutational catalogues and estimate the exposure of each
catalogue to each signature.
}
\examples{
# Load example mutational catalogues
data("counts_21breast")

# Fetch COSMIC signatures 
signatures <- fetch_cosmic_data()

# Fit signatures 1 to 4, using a custom prior that favors signature 1 over the rest
# (4 chains, 300 warmup iterations + 300 sampling iterations -- use more in practice)
samples_1 <- fit_signatures(counts_21breast, signatures[1:4, ], 
                            exp_prior = c(10, 1, 1, 1), iter = 600)

# Fit all the signatures, running a single chain for many iterations
# (3000 warmup iterations + 10000 sampling iterations)
samples_2 <- fit_signatures(counts_21breast, signatures, chains = 1, 
                            iter = 13000, warmup = 3000)
}
