% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_estimation.R
\name{fit_signatures}
\alias{fit_signatures}
\title{Fit mutational signatures}
\usage{
fit_signatures(counts, signatures, exp_prior = NULL, model = "nmf",
  opportunities = NULL, ...)
}
\arguments{
\item{counts}{Integer matrix of observed mutation counts, with one row per sample and 
one column per mutation type.}

\item{signatures}{Mutational signatures to be fitted. Either a numeric matrix with one row per signature
and one column per mutation type, or a list of matrices generated via
\code{\link{retrieve_pars}}.}

\item{exp_prior}{Numeric vector with one element per signature, to be used as the Dirichlet prior for 
the signature exposures in the sampling chain. Default prior is uniform (uninformative).}

\item{model}{Character; model to sample from. Admits values \code{"nmf"} (default) or \code{"emu"}.}

\item{opportunities}{Numeric matrix of optional mutational opportunities for the "EMu" model 
(\code{model = "emu"}). Must be a matrix with same dimension as \code{counts}. 
Alternatively, it also admits character values \code{"human-genome"} or \code{"human-exome"}, 
in which case the reference human genome/exome opportunities will be used for every sample.}

\item{...}{Arguments to pass to \code{rstan::sampling}.}
}
\value{
A list with two elements:
\itemize{
 \item{\code{$data}: list containing the input data supplied to the model.}
 \item{\code{$result}: object of class stanfit, containing the output MCMC samples, 
 as well as information about the model and sampling process.}}
The model parameters (such as signatures and exposures) can be extracted from this 
object using \code{\link{retrieve_parameters}}.
}
\description{
\code{fit_signatures} runs MCMC sampling to fit a set of mutational signatures 
to a collection of mutational catalogues and estimate the exposure of each
catalogue to each signature.
}
\examples{
# Load example mutational catalogues and COSMIC signatures
data("counts_21breast")
data("cosmic_signatures")

# Fit signatures 1 to 4, using a custom prior that favors signature 1 over the rest
# (4 chains, 300 warmup iterations + 300 sampling iterations - use more in practice)
samples_1 <- fit_signatures(counts_21breast, cosmic_signatures[1:4, ], 
                            exp_prior = c(10, 1, 1, 1), iter = 600)

# Fit all the signatures, running a single chain for many iterations
# (3000 warmup iterations + 10000 sampling iterations)
samples_2 <- fit_signatures(counts_21breast, cosmic_signatures, chains = 1, 
                            iter = 13000, warmup = 3000)
}
