% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{get_loglik}
\alias{get_loglik}
\title{Generate log likelihood values from a model}
\usage{
get_loglik(mcmc_samples)
}
\arguments{
\item{mcmc_samples}{List with elements \code{`data`} and \code{`results`}, produced via either
\code{\link{fit_signatures}}, \code{\link{extract_signatures}} or \code{\link{fit_extract_signatures}}.}
}
\description{
Generate log likelihood values from a model
}
