% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{human_trinuc_freqs}
\alias{human_trinuc_freqs}
\title{Retrieve human trinucleotide frequencies}
\usage{
human_trinuc_freqs(type = "human-genome", strand = FALSE)
}
\arguments{
\item{type}{Character; either \code{"genome"} (default) or \code{"exome"}.}

\item{strand}{Logical; if \code{TRUE}, a strand-bias representation of catalogues
and signatures will be used.}
}
\value{
A numeric vector containinig 96 frequency values (one per trinucleotide type), if
\code{strand=FALSE}, or 192 frequency values (one per trinucleotide and strand type), if
\code{strand=TRUE}. In the latter case, the trinucleotide frequencies are assumed to be
equally distributed between the two strands.
}
\description{
\code{human_trinuc_freqs} returns the reference human genome or exome 
trinucleotide frequencies.
}
