% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{match_signatures}
\alias{match_signatures}
\title{Find the best matches between two sets of signatures}
\usage{
match_signatures(sigs_a, sigs_b, stat = "cosine")
}
\arguments{
\item{sigs_a}{First set of signatures; either a numeric matrix of signatures,
with one row per signature and one column per mutation type, or a list of matrices
generated via \code{\link{retrieve_pars}}.}

\item{sigs_b}{Second set of signatures, as for \code{sigs_a}.}

\item{stat}{Similarity metric to use when comparing signatures. Admits values \code{"cosine"}
(default, cosine similarity) or \code{"L2"} (L2 norm or Euclidean distance).}
}
\value{
A numeric vector containing, for each signature in \code{sigs_a}, the index
of the closest match in \code{sigs_b}.
}
\description{
\code{match_signatures} compares two independent estimates of signatures to
find the closest matches between them.
}
