% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_plotting.R
\name{plot_all}
\alias{plot_all}
\title{Plot all results from signature fitting or extraction}
\usage{
plot_all(counts, out_path, prefix = NULL, mcmc_samples = NULL,
  signatures = NULL, exposures = NULL, opportunities = NULL,
  thresh = 0.01, horiz_labels = FALSE, hpd_prob = 0.95,
  signature_names = NULL, exp_legend_pos = "topright",
  rec_legend_pos = "topright", sig_color_palette = NULL)
}
\arguments{
\item{counts}{Matrix of observed mutation counts (integers), with one row per sample and 
column for each of the 96 mutation types.}

\item{out_path}{Path to the directory where the output PDF files will be stored. M
Will be created if it does not exist.}

\item{prefix}{Optional prefix to be added to the output file names.}

\item{mcmc_samples}{Object of class stanfit, generated via either \code{\link{fit_signatures}}
or \code{\link{extract_signatures}}. Only needed if \code{signatures} or \code{exposures}
are not provided.}

\item{signatures}{Either a matrix of mutational signatures, with one row per signature and one
column for each of the 96 mutation types, or a list of signatures generated via
\code{\link{retrieve_pars}}. Only needed if \code{mcmc_samples} is not provided, or if it
contains signature fitting (instead of signature extraction) results.}

\item{exposures}{Either a matrix of signature exposures, with one row per sample and one column 
per signature, or a list of exposures as produced by \code{\link{retrieve_pars}}. 
Only needed if \code{mcmc_samples} is not provided.}

\item{opportunities}{If \code{signatures} and/or \code{exposures} were obtained by extracting or fitting
signatures using the "EMu" model (\code{method = "emu"}), these should be the same opportunities used 
for extraction/fitting. Admits values \code{"human-genome"} and \code{"human-exome"}.}

\item{thresh}{Minimum probability value that should be reached by the lower end of exposure HPD
intervals. The exposures for which the lower HPD bound is below this value will be colored in grey.
This value is passed to the \code{plot_exposures} function.}

\item{horiz_labels}{If \code{TRUE}, sample name labels will be displayed horizontally in the
barplots. This value is passed to the \code{plot_exposures} function.}

\item{hpd_prob}{A value in the interval (0, 1), giving the target probability content of 
the HPD intervals. This value is passed to the \code{plot_exposures} function.}

\item{exp_legend_pos}{Character indicating the position of the legend in the exposures barplot. Admits values
\code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"}, 
\code{"topright"}, \code{"right"} and \code{"center"}. This value is passed to the \code{plot_exposures} function.}

\item{rec_legend_pos}{Character indicating the position of the legend in the reconstructed spectrum. Admits values
\code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"}, 
\code{"topright"}, \code{"right"} and \code{"center"}. This value is passed to the \code{plot_reconstruction} function.}

\item{sig_color_palette}{Character vector of color names or hexadecimal codes to use for each signature.
Must have at least as many elements as the number of signatures. This value is passed to the 
\code{plot_exposures} and \code{plot_reconstruction} functions.}
}
\description{
For a given set of signature fitting or extraction results, \code{plot_all} plots, in PDF format: 
\itemize{
 \item{All the original (input) mutational catalogues (via \code{\link{plot_spectrum}})}
 \item{Mutational signatures (via \code{\link{plot_spectrum}})}
 \item{Signature exposures (via \code{\link{plot_exposures}})}
 \item{All the reconstructed mutational spectra (via \code{\link{plot_reconstruction}})}
}
}
\examples{
# Load example mutational catalogues
data("counts_21breast")

# Extract signatures using the EMu (Poisson) model
samples <- extract_signatures(counts_21breast, nsignatures = 2, method = "emu",
                              opportunities = "human-genome", iter = 800)

# Retrieve signatures and exposures
signatures <- retrieve_pars(samples, "signatures")
exposures <- retrieve_pars(samples, "exposures")

# Plot results using stanfit object
plot_all(counts_21breast, out_path = ".", prefix = "Test1", 
         mcmc_samples = samples, opportunities = "human-genome")
                    
# Plot results using retrieved signatures and exposures
plot_all(counts_21breast, out_path = ".", prefix = "Test2", 
         signatures = signatures, exposures = exposures, 
         opportunities = "human-genome")
}
