% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_plotting.R
\name{plot_all}
\alias{plot_all}
\title{Plot all results from signature fitting or extraction}
\usage{
plot_all(mcmc_samples = NULL, out_path, prefix = NULL, counts = NULL,
  signatures = NULL, exposures = NULL, opportunities = NULL,
  thresh = 0.01, hpd_prob = 0.95, signature_names = NULL,
  exp_margin_bottom = 10.5, exp_legend_pos = "topleft",
  exp_legend_cex = 2, exp_cex_names = 1.9,
  rec_legend_pos = "topleft", rec_legend_cex = 2,
  sig_color_palette = NULL, boxes = TRUE)
}
\arguments{
\item{mcmc_samples}{List with two elements named \code{`data`} and \code{`results`}, produced via
\code{\link{fit_signatures}}, \code{\link{extract_signatures}}, or
\code{\link{fit_extract_signatures}}. This is the preferred option for supplying data and
results, but can be replaced by the combination of arguments \code{counts}, \code{signatures},
\code{exposures} and \code{opportunities}.}

\item{out_path}{Character indicating the path to the directory where the output PDF files will
be stored. The directory will be created if it does not exist.}

\item{prefix}{Character indicating an optional prefix to be added to output file names.}

\item{counts}{Numeric matrix of observed mutation counts, with one row per sample and
one column per mutation type. Only needed if \code{mcmc_samples} is not provided.}

\item{signatures}{Either a numeric matrix of mutational signatures, with one row per signature
and one column per mutation type, or a list of matrices generated via \code{\link{retrieve_pars}}.
Only needed if \code{mcmc_samples} is not provided.}

\item{exposures}{Either a numeric matrix of signature exposures, with one row per sample and one
column per signature, or a list of matrices generated via \code{\link{retrieve_pars}}. Only
needed if \code{mcmc_samples} is not provided.}

\item{opportunities}{Numeric matrix of mutational opportunities, with one row per signature and
one column per mutation type. It also admits character values \code{"human-genome"} or
\code{"human-exome"}, in which case the mutational opportunities of the reference human
genome/exome will be used. Only needed if \code{mcmc_samples} is not provided and opportunities
were used during signature extraction or fitting.}

\item{thresh}{Numeric indicating the minimum threshold for the lower HPD limits of signature
exposures (default is 0.01). Exposures with a lower HPD limit below this value will be shown in
grey. This value is passed to \code{\link{plot_exposures}}.}

\item{hpd_prob}{Numeric value in the interval (0, 1), indicating the desired probability content
of HPD intervals (default is 0.95). This value is passed to \code{\link{plot_exposures}}.}

\item{signature_names}{Character vector containing the name of each signature. Only needed if
\code{mcmc_samples} is not provided and the exposures were obtained via signature fitting
(rather than extraction). This value is passed to \code{\link{plot_exposures}}.}

\item{exp_margin_bottom}{Numeric indicating the bottom margin of the exposures barplots, in
inches (default is 10.5). This value is passed to \code{\link{plot_exposures}}.}

\item{exp_legend_pos}{Character indicating the position of the legend in the exposures barplot.
Admits values \code{"top"}, \code{"bottom"}, \code{"center"}, \code{"left"}, \code{"right"}, 
\code{"topleft"}, \code{"topright"}, \code{"bottomleft"} and \code{"bottomright"} (default is
\code{"topleft"}). This value is passed to \code{\link{plot_exposures}}.}

\item{exp_legend_cex}{Numeric indicating the relative size of the legend in the exposures
barplot (default is 2). This value is passed to \code{\link{plot_exposures}}.}

\item{exp_cex_names}{Numeric indicating the relative size of sample labels in the exposures
barplot (default is 1.9). This value is passed to \code{\link{plot_exposures}}.}

\item{rec_legend_pos}{Character indicating the position of the legend in the spectrum
reconstruction plots. Admits values \code{"top"}, \code{"bottom"}, \code{"center"}, \code{"left"},
\code{"right"}, \code{"topleft"}, \code{"topright"}, \code{"bottomleft"} and \code{"bottomright"}
(default is \code{"topleft"}). This value is passed to \code{\link{plot_reconstruction}}.}

\item{rec_legend_cex}{Numeric indicating the relative size of the legend in the reconstruction
plots (default is 2). This value is passed to \code{\link{plot_reconstruction}}.}

\item{sig_color_palette}{Character vector of custom color names or hexadecimal codes to use for
each signature in exposure and reconstruction plots. Must have at least as many elements as the
number of signatures. This value is passed to \code{\link{plot_exposures}} and
\code{\link{plot_reconstruction}}.}

\item{boxes}{Logical indicating whether boxes should be drawn around spectrum, signature and
reconstruction plots (default is \code{TRUE}). This value is passed to
\code{\link{plot_spectrum}} and \code{\link{plot_reconstruction}}.}
}
\description{
For a given set of signature fitting or extraction results, \code{plot_all} plots, in PDF format:
\itemize{
 \item{All the original (input) mutational catalogues (via \code{\link{plot_spectrum}})}
 \item{Mutational signatures (via \code{\link{plot_spectrum}})}
 \item{Signature exposures (via \code{\link{plot_exposures}})}
 \item{All the reconstructed mutational spectra (via \code{\link{plot_reconstruction}})}
}
}
\examples{
\dontrun{
# Load example mutational catalogues
data("counts_21breast")

# Extract signatures using the EMu (Poisson) model
samples <- extract_signatures(counts_21breast, nsignatures = 2, model = "emu",
                              opportunities = "human-genome", iter = 800)

# Retrieve signatures and exposures
signatures <- retrieve_pars(samples, "signatures")
exposures <- retrieve_pars(samples, "exposures")

# Plot results using MCMC samples
plot_all(mcmc_samples = samples, out_path = ".", prefix = "Test1")

# Plot results using retrieved signatures and exposures
plot_all(counts = counts_21breast, signatures = signatures,
         exposures = exposures, opportunities = "human-genome",
         out_path = ".", prefix = "Test2")
}
}
