% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_plotting.R
\name{plot_exposures}
\alias{plot_exposures}
\title{Plot signature exposures}
\usage{
plot_exposures(mcmc_samples = NULL, pdf_path = NULL, counts = NULL,
  exposures = NULL, signature_names = NULL, thresh = 0.01,
  hpd_prob = 0.95, horiz_labels = FALSE, legend_pos = "topright",
  sig_color_palette = NULL)
}
\arguments{
\item{mcmc_samples}{List with elements \code{$data} and \code{$results}, produced via either 
\code{\link{fit_signatures}}, \code{\link{extract_signatures}} or \code{\link{fit_extract_signatures}}. 
This should be the preferred option for supplying data and results, but can be replaced by the 
combination of arguments \code{counts}, \code{exposures} and \code{signature_names}.}

\item{pdf_path}{Character; if provided, the plots will be output to a PDF file with this path. The PDF 
size and graphical parameters will be automatically set to appropriate values.}

\item{counts}{Integer matrix of observed mutation counts, with one row per sample and 
column for each of the 96 (or 192) mutation types. Only needed if \code{mcmc_samples} is not provided.}

\item{exposures}{Either a numeric matrix of signature exposures, with one row per sample and one column 
per signature, or a list of exposures as produced by \code{\link{retrieve_pars}}. 
Only needed if \code{mcmc_samples} is not provided.}

\item{signature_names}{Character vector containing the signature names. Only needed if \code{mcmc_samples}
is not provided and the exposures were obtained through signature fitting (not extraction).}

\item{thresh}{Numeric; minimum probability value that should be reached by the lower end of exposure HPD
intervals. The exposures for which the lower HPD bound is below this value will be colored in grey.
(Default is 0.01.)}

\item{hpd_prob}{Numeric; a value in the interval (0, 1), giving the target probability content of 
the HPD intervals. (Default is 0.95.)}

\item{horiz_labels}{Logical; if \code{TRUE}, sample name labels will be displayed horizontally in the
barplots. (Default is \code{FALSE}.)}

\item{legend_pos}{Character indicating the position of the legend in the exposures barplot. Admits values
\code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"}, 
\code{"topright"} (default), \code{"right"} and \code{"center"}.}

\item{sig_color_palette}{Character vector of color names or hexadecimal codes to use for each signature.
Must have at least as many elements as the number of signatures.}
}
\description{
\code{plot_exposures} plots the distribution of signature exposures across the samples.
}
