% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{plot_exposures}
\alias{plot_exposures}
\title{Plots estimated exposure of each signature}
\usage{
plot_exposures(samples, prob = 0.9, thresh = 0.001,
  title = "Signature exposures")
}
\description{
Plots estimated exposure of each signature
}

