% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{plot_exposures}
\alias{plot_exposures}
\title{Plot signature exposures
\code{plot_exposures} produces barplots that show the distribution of
signatures exposures across catalogues.}
\usage{
plot_exposures(counts, exposures = NULL, mcmc_samples = NULL,
  pdf_path = NULL, sig_color_palette = NULL)
}
\description{
Plot signature exposures
\code{plot_exposures} produces barplots that show the distribution of
signatures exposures across catalogues.
}
