% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_plotting.R
\name{plot_exposures}
\alias{plot_exposures}
\title{Plot signature exposures}
\usage{
plot_exposures(counts, exposures = NULL, mcmc_samples = NULL,
  pdf_path = NULL, signature_names = NULL, thresh = 0.01,
  hpd_prob = 0.95, horiz_labels = FALSE, legend_pos = "topright",
  sig_color_palette = NULL)
}
\arguments{
\item{counts}{Matrix of observed mutation counts (integers), with one row per sample and 
column for each of the 96 mutation types.}

\item{exposures}{Either a matrix of signature exposures, with one row per sample and one column 
per signature, or a list of exposures as produced by \code{\link{retrieve_pars}}. 
Only needed if \code{mcmc_samples} is not provided.}

\item{mcmc_samples}{Object of class stanfit, generated via either \code{\link{fit_signatures}}
or \code{\link{extract_signatures}}. Only needed if \code{exposures} is not provided.}

\item{pdf_path}{If provided, the plots will be output to a PDF file with this path. The PDF 
size and graphical parameters will be automatically set to appropriate values.}

\item{signature_names}{Vector containing the names of the signatures. Only used when plotting
exposures obtained through signature fitting (not extraction).}

\item{thresh}{Minimum probability value that should be reached by the lower end of exposure HPD
intervals. The exposures for which the lower HPD bound is below this value will be colored in grey.}

\item{hpd_prob}{A value in the interval (0, 1), giving the target probability content of 
the HPD intervals.}

\item{horiz_labels}{If \code{TRUE}, sample name labels will be displayed horizontally in the
barplots.}

\item{legend_pos}{Character indicating the position of the legend in the exposures barplot. Admits values
\code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"}, 
\code{"topright"}, \code{"right"} and \code{"center"}.}

\item{sig_color_palette}{Character vector of color names or hexadecimal codes to use for each signature.
Must have at least as many elements as the number of signatures.}
}
\description{
\code{plot_exposures} plots the distribution of signature exposures across the samples.
}
