% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_plotting.R
\name{plot_gof}
\alias{plot_gof}
\title{Plot goodness of fit}
\usage{
plot_gof(sample_list, stat = "cosine")
}
\arguments{
\item{sample_list}{List with elements \code{$data} and \code{$results}, produced via 
\code{\link{extract_signatures}}. Must contain the results of signature extraction using
multiple numbers of signatures; i.e. the \code{nsignature} argument must have contained
multiple values when \code{\link{extract_signatures}} was run.}

\item{stat}{Character; function for measuring goodness of fit. Admits values \code{"cosine"} 
(default, cosine similarity) or \code{"L2"} (L2 norm or Euclidean distance).}
}
\description{
\code{plot_gof} plots the goodness of fit of a set of samples, each of which
has typically been sampled using an extraction model (EMu or NMF) with a 
different number of signatures.
}
