% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_plotting.R
\name{plot_gof}
\alias{plot_gof}
\title{Plot goodness of fit}
\usage{
plot_gof(sample_list, counts, stat = "cosine")
}
\arguments{
\item{sample_list}{List of objects of class stanfit. Elements which are not of class stanfit are ignored.}

\item{counts}{Integer matrix of observed mutation counts, with one row per sample and 
column for each of the 96 mutation types.}

\item{stat}{Character, function for measuring goodness of fit. Admits values \code{"cosine"} 
(cosine similarity; default) or \code{"L2"} (L2 norm, a.k.a. Euclidean distance).}
}
\description{
\code{plot_gof} plots the goodness of fit of a set of samples, each of which
has typically been sampled using an extraction model (EMu or NMF) with a 
different number of signatures.
}
