% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_plotting.R
\name{plot_reconstruction}
\alias{plot_reconstruction}
\title{Plot mutational spectrum reconstructions}
\usage{
plot_reconstruction(counts, mcmc_samples = NULL, signatures = NULL,
  exposures = NULL, opportunities = NULL, pdf_path = NULL,
  legend_pos = "topright", sig_color_palette = NULL)
}
\arguments{
\item{counts}{Matrix of observed mutation counts (integers), with one row per sample and 
column for each of the 96 mutation types.}

\item{mcmc_samples}{Object of class stanfit, generated via either \code{\link{fit_signatures}}
or \code{\link{extract_signatures}}. Only needed if \code{signatures} or \code{exposures}
are not provided.}

\item{signatures}{Either a matrix of mutational signatures, with one row per signature and one
column for each of the 96 mutation types, or a list of signatures generated via
\code{\link{retrieve_pars}}. Only needed if \code{mcmc_samples} is not provided.}

\item{exposures}{Either a matrix of signature exposures, with one row per sample and one column 
per signature, or a list of exposures as produced by \code{\link{retrieve_pars}}. 
Only needed if \code{mcmc_samples} is not provided.}

\item{opportunities}{If \code{signatures} and/or \code{exposures} were obtained by extracting or fitting
signatures using the "EMu" model (\code{method = "emu"}), these should be the same opportunities used 
for extraction/fitting. Admits values \code{"human-genome"} and \code{"human-exome"}.}

\item{pdf_path}{If provided, the plots will be output to a PDF file with this path. The PDF 
size and graphical parameters will be automatically set to appropriate values.}

\item{legend_pos}{Character indicating the position of the legend in the reconstructed spectrum. Admits values
\code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"}, 
\code{"topright"}, \code{"right"} and \code{"center"}.}

\item{sig_color_palette}{Character vector of color names or hexadecimal codes to use for each signature.
Must have at least as many elements as the number of signatures.}
}
\description{
\code{plot_reconstruction} plots the reconstructions of the original mutational catalogues using the 
signatures and/or exposures obtained through extraction or fitting. If provided with multiple catalogues, 
it generates one plot per catalogue. Fitting or extraction results can be provided either as a single 
stanfit object (generated via \code{\link{fit_signatures}} or \code{\link{extract_signatures}}), 
or as separate signatures and exposures matrices (or lists produced via \code{\link{retrieve_pars}}). 
Only the former option allows the incorporation of HPD intervals to the reconstructed catalogue.
}
\examples{
# Load example mutational catalogues
data("counts_21breast")

# Extract signatures using the EMu (Poisson) model
samples <- extract_signatures(counts_21breast, nsignatures = 2, method = "emu",
                              opportunities = "human-genome", iter = 800)

# Retrieve signatures and exposures
signatures <- retrieve_pars(samples, "signatures")
exposures <- retrieve_pars(samples, "exposures")

# Plot reconstructed catalogues using stanfit object
plot_reconstruction(counts_21breast, mcmc_samples = samples, opportunities = "human-genome",
                    pdf_path = "Reconstructions_1.pdf")
                    
# Plot reconstructed catalogues using retrieved signatures and exposures
plot_reconstruction(counts_21breast, signatures = signatures, exposures = exposures,
                    opportunities = "human-genome", pdf_path = "Reconstructions_2.pdf")
}
