% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot mutational spectra}
\usage{
plot_spectrum(spectra, name = NULL, pdf_path = NULL, max_y = NULL)
}
\arguments{
\item{spectra}{Either a vector with one element for each of the 96 mutation types, or a matrix 
with 96 columns and one row per signature/catalogue, or a list of signatures as produced by 
\code{\link{retrieve_pars}}. In the latter case, HPD intervals will also be plotted. 
Row names will be adopted as the sample/signature names.}

\item{name}{Name to include in the plot title; useful when plotting a single spectrum.}

\item{pdf_path}{If provided, the plots will be output to a PDF file with this path. The PDF 
size and graphical parameters will be automatically set to appropriate values.}

\item{max_y}{Fixed maximum limit of the y-axis (if necessary).}
}
\description{
\code{plot_spectrum} generates plots of one or more spectra, which can be either mutational 
catalogues or mutational signatures. If the spectra contain values above 1, the values will be 
interpreted as mutation counts (as in a catalogue); otherwise, they will be interpreted as 
mutation probabilities (as in a signature). If multiple spectra are provided, one plot per 
spectrum is produced.
}
\examples{
# Load example mutational catalogues
data("counts_21breast")

# Plot catalogues
plot_spectrum(counts_21breast, pdf_path = "Catalogues.pdf")

# Extract signatures using the EMu (Poisson) model
samples <- extract_signatures(counts_21breast, nsignatures = 2, method = "emu",
                              opportunities = "human-genome", iter = 800)

# Retrieve extracted signatures
sigs <- retrieve_pars(samples, "signatures")

# Plot signatures
plot_spectrum(sigs, pdf_path = "Signatures.pdf")
}
