% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit.R
\name{retrieve_pars}
\alias{retrieve_pars}
\title{Retrieve model parameters}
\usage{
retrieve_pars(mcmc_samples, feature, hpd_prob = 0.95,
  signature_names = NULL)
}
\arguments{
\item{mcmc_samples}{An object of class stanfit, resulting from MCMC sampling through either \code{fit_signatures} or \code{extract_signatures}.}

\item{feature}{Name of the parameter set to extract; either \code{"signatures"} or \code{"exposures"}.}

\item{hpd_prob}{A numeric scalar in the interval (0,1) giving the target probability content of the HPD intervals.}

\item{signature_names}{Vector containing the names of the signatures used for fitting. Used only when 
retrieving exposures from fitted signatures.}
}
\description{
Obtains summary values for a set of model parameters (signatures or exposures) from a stanfit object.
}
\examples{
# Extract signatures using the EMu (Poisson) model
samples <- extract_signatures(mycounts, nsignatures = 3, method = "emu", 
opportunities = "human-genome")

# Retrieve array of signatures
signatures <- retrieve_pars(samples, "signatures")

# Retrieve array of exposures using 90\% HPD intervals
exposures <- retrieve_pars(samples, "exposures", hpd_prob = 0.9)

# Plot signatures and mean exposures
plot_spectrum(signatures)
barplot(exposures$mean)   ## or barplot(exposures[[1]])
}
