% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{retrieve_pars}
\alias{retrieve_pars}
\title{Retrieve model parameters}
\usage{
retrieve_pars(mcmc_samples, feature, hpd_prob = 0.95, counts = NULL,
  signatures = NULL, signature_names = NULL)
}
\arguments{
\item{mcmc_samples}{Object of class stanfit, generated via either \code{\link{fit_signatures}}
or \code{\link{extract_signatures}}.}

\item{feature}{Character, name of the parameter set to extract. Can take values: \code{"signatures"}, 
\code{"exposures"}, \code{"activities"} and \code{"reconstructions"}.}

\item{hpd_prob}{Numeric, a value in the interval (0, 1), giving the target probability content of 
the HPD intervals.}

\item{signature_names}{Character vector containing the names of the signatures used for fitting. Used only when 
retrieving exposures from fitted signatures.}
}
\value{
A list of three matrices, which respectively contain the values corresponding to the
mean of the model parameter of interest, and to the lower and upper ends of its HPD interval.
}
\description{
Obtains summary values for a set of model parameters (signatures or exposures) from a stanfit object.
}
\examples{
# Load example mutational catalogues
data("counts_21breast")

# Extract signatures using the EMu (Poisson) model
samples <- extract_signatures(counts_21breast, nsignatures = 2, method = "emu",
                              opportunities = "human-genome", iter = 800)

# Retrieve signatures
signatures <- retrieve_pars(samples, "signatures")

# Retrieve exposures and activities using custom HPD intervals
exposures <- retrieve_pars(samples, "exposures", hpd_prob = 0.9)
activities <- retrieve_pars(samples, "activities", hpd_prob = 0.975)

# Retrieve reconstructed catalogues (reconstructions)
reconstructions <- retrieve_pars(samples, "reconstructions")

# Plot signatures, reconstructions and mean exposures
plot_spectrum(signatures)
plot_spectrum(reconstructions)
barplot(exposures$mean)
}
