% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{retrieve_pars}
\alias{retrieve_pars}
\title{Retrieve model parameters}
\usage{
retrieve_pars(mcmc_samples, feature, hpd_prob = 0.95,
  signature_names = NULL)
}
\arguments{
\item{mcmc_samples}{Object of class stanfit, generated via either \code{\link{fit_signatures}}
or \code{\link{extract_signatures}}.}

\item{feature}{Name of the parameter set to extract; either \code{"signatures"} or \code{"exposures"}.}

\item{hpd_prob}{A value in the interval (0, 1), giving the target probability content of 
the HPD intervals.}

\item{signature_names}{Vector containing the names of the signatures used for fitting. Used only when 
retrieving exposures from fitted signatures.}
}
\value{
A list of three matrices, which respectively contain the values corresponding to the
mean of the model parameter of interest, and to the lower and upper ends of its HPD interval.
}
\description{
Obtains summary values for a set of model parameters (signatures or exposures) from a stanfit object.
}
\examples{
# Load example mutational catalogues
data("counts_21breast")

# Extract signatures using the EMu (Poisson) model
samples <- extract_signatures(counts_21breast, nsignatures = 2, method = "emu",
                              opportunities = "human-genome", iter = 800)

# Retrieve signatures and exposures
signatures <- retrieve_pars(samples, "signatures")

# Retrieve array of exposures using 90\% HPD intervals
exposures <- retrieve_pars(samples, "exposures", hpd_prob = 0.9)

# Plot signatures
plot_spectrum(signatures)

# Plot mean exposures
barplot(exposures$mean)   ## or: barplot(exposures[[1]])
}
