% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{retrieve_pars}
\alias{retrieve_pars}
\title{Retrieve model parameters}
\usage{
retrieve_pars(mcmc_samples, par, hpd_prob = 0.95)
}
\arguments{
\item{mcmc_samples}{List with two elements named \code{`data`} and \code{`results`}, produced via
\code{\link{fit_signatures}}, \code{\link{extract_signatures}}, or
\code{\link{fit_extract_signatures}}.}

\item{par}{Name of the parameter set to extract. Admits character values \code{"signatures"},
\code{"exposures"}, \code{"activities"} or \code{"reconstructions"}.}

\item{hpd_prob}{Numeric value in the interval (0, 1), indicating the desired probability content
of HPD intervals (default is 0.95).}
}
\value{
A list of three matrices, which contain the values corresponding to the means of the
model parameters and to the lower and upper ends of their HPD intervals, respectively.
}
\description{
\code{retrieve_pars} obtains summary values for a set of model parameters (signatures, exposures,
activities or spectrum reconstructions) from a stanfit object.
}
\examples{
\dontrun{
# Load example mutational catalogues
data("counts_21breast")

# Extract signatures using the EMu (Poisson) model
samples <- extract_signatures(counts_21breast, nsignatures = 2, model = "emu",
                              opportunities = "human-genome", iter = 800)

# Retrieve signatures
signatures <- retrieve_pars(samples, "signatures")

# Retrieve exposures and activities using custom HPD intervals
exposures <- retrieve_pars(samples, "exposures", hpd_prob = 0.9)
activities <- retrieve_pars(samples, "activities", hpd_prob = 0.975)

# Retrieve reconstructed catalogues (reconstructions)
reconstructions <- retrieve_pars(samples, "reconstructions")

# Plot signatures, reconstructions and mean exposures
plot_spectrum(signatures)
plot_spectrum(reconstructions)
barplot(t(exposures$mean))
}
}
