% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{rev_comp}
\alias{rev_comp}
\title{Reverse complement a nucleotide sequence string
Input is string, output is character vector}
\usage{
rev_comp(nucleotides)
}
\description{
Reverse complement a nucleotide sequence string
Input is string, output is character vector
}
