% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{simulate_ppc}
\alias{simulate_ppc}
\title{Generate posterior predictive check values from a model}
\usage{
simulate_ppc(counts, mcmc_samples)
}
\arguments{
\item{counts}{Integer matrix of observed mutation counts, with one row per sample and 
column for each of the 96 mutation types.}

\item{mcmc_samples}{Object of class stanfit, generated via either \code{\link{fit_signatures}}
or \code{\link{extract_signatures}}.}
}
\description{
Generate posterior predictive check values from a model
}
