% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfit_utility.R
\name{simulate_ppc}
\alias{simulate_ppc}
\title{Generate posterior predictive check values from a model}
\usage{
simulate_ppc(mcmc_samples)
}
\arguments{
\item{mcmc_samples}{List with two elements named \code{`data`} and \code{`results`}, produced via
\code{\link{fit_signatures}}, \code{\link{extract_signatures}}, or
\code{\link{fit_extract_signatures}}.}
}
\description{
Generate posterior predictive check values from a model
}
